+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # 0.2.0 [Unreleased]
 
+- Add backgrounds collection for Christmas.
 - Add option to view the latest changelog.
 - Auto show changelog when a new version is installed.
 - Add `Liked photos` dialog to view all your liked photos.
 - Fix version crash.
 - Improve performance for dropdowns.
 - Add reset settings option.
-- Add backgrounds collection for Christmas.
+- Add options for donation, sponsor, and issue report.
 
 # 0.1.1